OFZIPArchiveEntry: Make (un)compressedSize ull
This is for consistency with OFTarArchiveEntry and OFLHAArchiveEntry.

FossilOrigin-Name: 94c326ab78f32a9cfb684d4547989c42b2fa6fce79b96602405802c6b0d22b8a

+++ src/OFMutableZIPArchiveEntry.h
@@ -92,12 +92,12 @@ OF_ASSUME_NONNULL_BEGIN
 /**
  * @brief The compressed size of the entry's file.
  */
-@property (readwrite, nonatomic) uint64_t compressedSize;
+@property (readwrite, nonatomic) unsigned long long compressedSize;
 
 /**
  * @brief The uncompressed size of the entry's file.
  */
-@property (readwrite, nonatomic) uint64_t uncompressedSize;
+@property (readwrite, nonatomic) unsigned long long uncompressedSize;
 
 /**
  * @brief The CRC32 checksum of the entry's file.
